+++ library/src/dapp-connection.js
@@ -1,14 +1,14 @@
 const ParentStream = require('iframe-stream').ParentStream
-const SWcontroller = require('./sw-controller')
+const SWcontroller = require('client-sw-ready-event/lib/sw-client.js')
 const SwStream = require('sw-stream/lib/sw-stream.js')
 const SetupUntrustedComunication = ('./lib/setup-untrusted-connection.js')
+
 const background = new SWcontroller({
   fileName: '/popup/sw-build.js',
 })
 
 const pageStream = new ParentStream()
 background.on('ready', (_) => {
-  // var inpageProvider = new MetamaskInpageProvider(SwStream(background.controller))
   let swStream = SwStream({
     serviceWorker: background.controller,
     context: 'dapp',
