+++ package.json
@@ -42,6 +42,7 @@
     "bip39": "^2.2.0",
     "browser-passworder": "^2.0.3",
     "browserify-derequire": "^0.9.4",
+    "client-sw-ready-event": "^1.0.2",
     "clone": "^1.0.2",
     "copy-to-clipboard": "^2.0.0",
     "debounce": "^1.0.0",
